Move code out of data_provider
CMK-12002
